create file of constants
