+++ src/config/constants.tsx
@@ -1,0 +1,12 @@
+export const INPUT_PHONE_MASK = '(99) 9 9999-9999'
+
+export const PROFILES: Record<string, { label: string; value: string }> = {
+  ADMIN: {
+    label: "admin",
+    value: "admin",
+  },
+  USER: {
+    label: "user",
+    value: "user",
+  },
+};
